refactor: disconnecting interface from the page
