+++ app/(main)/projects/page.tsx
@@ -2,19 +2,7 @@
 import { notFound } from 'next/navigation'
 
 import { getPages } from '@/libs/shared/helpers/notion.helpers'
-
-interface PartialPageObjectResponseMore {
-  object: 'page'
-  id: string
-  properties?: {
-    Slug: {
-      rich_text: Array<{ plain_text: string }>
-    }
-    Title: {
-      title: Array<{ plain_text: string }>
-    }
-  }
-}
+import { PartialPageObjectResponseMore } from '@/libs/shared/types/page.type'
 
 export default async function ProjectsPage({
   searchParams,
