+++ libs/shared/types/page.type.ts
@@ -1,0 +1,14 @@
+interface PartialPageObjectResponseMore {
+  object: 'page'
+  id: string
+  properties?: {
+    Slug: {
+      rich_text: Array<{ plain_text: string }>
+    }
+    Title: {
+      title: Array<{ plain_text: string }>
+    }
+  }
+}
+
+export type { PartialPageObjectResponseMore }
